+++ service-base/src/main/java/org/eclipse/hono/service/management/credentials/GenericSecret.java
@@ -35,5 +35,4 @@ public class GenericSecret extends CommonSecret {
     public Map<String, Object> getAdditionalProperties() {
         return this.additionalProperties;
     }
-
 }